+++ action.go
@@ -11,8 +11,6 @@ import (
 	"strings"
 	"sync"
 
-	"golang.org/x/xerrors"
-
 	"github.com/mediocregopher/radix/v3/resp"
 	"github.com/mediocregopher/radix/v3/resp/resp2"
 )
@@ -470,13 +468,13 @@ func (p pipeline) drain(c Conn, n int) {
 func decodeErr(cmd CmdAction, err error) error {
 	c, ok := cmd.(*cmdAction)
 	if ok {
-		return xerrors.Errorf(
+		return fmt.Errorf(
 			"failed to decode pipeline CmdAction '%v' with keys %v: %w",
 			c.cmd,
 			c.Keys(),
 			err)
 	}
-	return xerrors.Errorf(
+	return fmt.Errorf(
 		"failed to decode pipeline CmdAction '%v': %w",
 		cmd,
 		err)
@@ -500,7 +498,6 @@ func (p pipeline) MarshalRESP(w io.Writer) error {
 			return err
 		}
 	}
-
 	return nil
 }