+++ internal/bytesutil/bytesutil.go
@@ -4,13 +4,13 @@ package bytesutil
 
 import (
 	"bufio"
+	"errors"
 	"fmt"
 	"io"
 	"strconv"
 	"sync"
 
 	"github.com/mediocregopher/radix/v3/resp"
-	errors "golang.org/x/xerrors"
 )
 
 // AnyIntToInt64 converts a value of any of Go's integer types (signed and unsigned) into a signed int64.
@@ -102,7 +102,7 @@ func ParseUint(b []byte) (uint64, error) {
 
 	for i, c := range b {
 		if c < '0' || c > '9' {
-			return 0, errors.Errorf("invalid character %c at position %d in parseUint", c, i)
+			return 0, fmt.Errorf("invalid character %c at position %d in parseUint", c, i)
 		}
 
 		n *= 10
@@ -130,7 +130,7 @@ func BufferedBytesDelim(br *bufio.Reader) ([]byte, error) {
 	if err != nil {
 		return nil, err
 	} else if len(b) < 2 || b[len(b)-2] != '\r' {
-		return nil, errors.Errorf("malformed resp %q", b)
+		return nil, fmt.Errorf("malformed resp %q", b)
 	}
 	return b[:len(b)-2], err
 }