+++ pubsub.go
@@ -3,12 +3,11 @@ package radix
 import (
 	"bufio"
 	"bytes"
+	"errors"
 	"io"
 	"net"
 	"sync"
 	"time"
-
-	errors "golang.org/x/xerrors"
 
 	"github.com/mediocregopher/radix/v3/resp"
 	"github.com/mediocregopher/radix/v3/resp/resp2"