+++ radix.go
@@ -171,7 +171,7 @@ import (
 	"strings"
 	"time"
 
-	errors "golang.org/x/xerrors"
+	"errors"
 
 	"github.com/mediocregopher/radix/v3/resp"
 )