+++ sentinel.go
@@ -1,12 +1,11 @@
 package radix
 
 import (
+	"fmt"
 	"net"
 	"sync"
 	"sync/atomic"
 	"time"
-
-	errors "golang.org/x/xerrors"
 )
 
 type sentinelOpts struct {
@@ -338,7 +337,7 @@ func (sc *Sentinel) Close() error {
 // cmd should be the command called which generated m
 func sentinelMtoAddr(m map[string]string, cmd string) (string, error) {
 	if m["ip"] == "" || m["port"] == "" {
-		return "", errors.Errorf("malformed %q response: %#v", cmd, m)
+		return "", fmt.Errorf("malformed %q response: %#v", cmd, m)
 	}
 	return net.JoinHostPort(m["ip"], m["port"]), nil
 }