+++ stub_test.go
@@ -8,8 +8,6 @@ import (
 	. "testing"
 	"time"
 
-	errors "golang.org/x/xerrors"
-
 	"github.com/mediocregopher/radix/v3/resp/resp2"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -29,7 +27,7 @@ func testStub() Conn {
 		case "ECHO":
 			return args[1]
 		default:
-			return errors.Errorf("testStub doesn't support command %q", args[0])
+			return fmt.Errorf("testStub doesn't support command %q", args[0])
 		}
 	})
 }
@@ -115,7 +113,7 @@ func ExampleStub() {
 			m[args[1]] = args[2]
 			return nil
 		default:
-			return errors.Errorf("this stub doesn't support command %q", args[0])
+			return fmt.Errorf("this stub doesn't support command %q", args[0])
 		}
 	})